requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
flask>=3.0.0
